requirements.txt
# ERC Trading System - Dependencies

# Core
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
aiohttp>=3.8.0
requests>=2.26.0

# API Clients
alpaca-trade-api>=2.0.0
python-dotenv>=0.19.0

# Web Framework (for ML engine API)
flask>=2.0.0

# ML/Statistics
scikit-learn>=1.0.0
xgboost>=1.5.0

# Semantic Matching (optional, improves cross-platform matching)
sentence-transformers>=2.2.0

# Database
# (uses built-in sqlite3)

# Testing
pytest>=7.0.0
pytest-asyncio>=0.18.0

# Development
black>=22.0.0
flake8>=4.0.0
